styled

+++ src/logs/LocationLog.js
@@ -1,9 +1,7 @@
 import React, { useState, useEffect } from "react";
-import EntryCard from "./entryCard";
 import Grid from "@material-ui/core/Grid";
 import InputLabel from "@material-ui/core/InputLabel";
 import Select from "@material-ui/core/Select";
-import { makeStyles } from "@material-ui/core/styles";
 import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
@@ -12,32 +10,34 @@ import TableHead from "@material-ui/core/TableHead";
 import TableRow from "@material-ui/core/TableRow";
 import Typography from "@material-ui/core/Typography";
 import Paper from "@material-ui/core/Paper";
+import DeleteOutlinedIcon from "@material-ui/icons/DeleteOutlined";
+import EditIcon from "@material-ui/icons/Edit";
 import apiManager from "../api/apiManager";
 
 const LocationLog = (props) => {
   const locations = props.locations;
+  // const routes = props.routes;
   const events = props.events;
   const dates = props.uniqueDates;
   const chosenLocation = props.chosenLocation;
+  const chosenRoute = props.chosenRoute;
   const chosenEvent = props.chosenEvent;
   const chosenDate = props.chosenDate;
   const handleChosenLocationChange = props.handleChosenLocationChange;
+  // const handleChosenRouteChange = props.handleChosenRouteChange;
   const handleChosenEventChange = props.handleChosenEventChange;
   const handleChosenDateChange = props.handleChosenDateChange;
   const [entries, setEntries] = useState([]);
 
-  //  get token for authentication
-  let token = window.sessionStorage.getItem("token");
-
-  // get entries based on location and event chosen from dropdowns then filter based on date
+  // get entries based on location and event chosen from dropdowns, filter by date, then sort by time
   const getEntries = (locationId, eventId) => {
     apiManager.getEntriesByLocationAndEvent(locationId, eventId).then((r) => {
       setEntries(r);
     });
   };
-  const entriesByDate = entries.filter((entry) =>
-    entry.date.includes(chosenDate)
-  );
+  const entriesByDate = entries
+    .filter((entry) => entry.date.includes(chosenDate))
+    .sort((a, b) => a.time.localeCompare(b.time));
 
   let totalAttendeeCount = 0;
   if (entriesByDate.length !== 0) {
@@ -46,10 +46,15 @@ const LocationLog = (props) => {
       .reduce((accumulator, runningTotal) => accumulator + runningTotal);
   }
 
-  const deleteThisEntry = (id, locationId, eventId) => {
-    apiManager.deleteEntry(id).then(() => {
-      getEntries(props.chosenLocation, props.chosenEvent);
-    });
+  const deleteThisEntry = (id) => {
+    const check = window.confirm(
+      "Are you sure you want to delete this entry? Deletion cannot be undone."
+    );
+    if (check == true) {
+      apiManager.deleteEntry(id).then(() => {
+        getEntries(props.chosenLocation, props.chosenEvent);
+      });
+    }
   };
 
   useEffect(() => {
@@ -77,14 +82,13 @@ const LocationLog = (props) => {
         </Typography>
         <Grid container spacing={3}>
           <Grid item xs={12} md={3}>
-            <InputLabel>Select Event:</InputLabel>
+            <InputLabel>Event:</InputLabel>
             <Select
               id="eventId"
               native
               onChange={handleChosenEventChange}
               fullWidth
               required
-              label="??"
               value={chosenEvent}
             >
               {events ? (
@@ -99,14 +103,13 @@ const LocationLog = (props) => {
             </Select>
           </Grid>
           <Grid item xs={12} md={3}>
-            <InputLabel>Select Location:</InputLabel>
+            <InputLabel>Location:</InputLabel>
             <Select
               id="locationId"
               native
               onChange={handleChosenLocationChange}
               fullWidth
               required
-              label="??"
               value={chosenLocation}
             >
               {locations ? (
@@ -120,15 +123,35 @@ const LocationLog = (props) => {
               )}
             </Select>
           </Grid>
+          {/* <Grid item xs={12} md={3}>
+            <InputLabel>Route:</InputLabel>
+            <Select
+              id="routeId"
+              native
+              onChange={handleChosenRouteChange}
+              fullWidth
+              required
+              value={chosenRoute}
+            >
+              {routes ? (
+                routes.map((route) => (
+                  <option key={route.id} value={parseInt(route.id)}>
+                    {route.name}
+                  </option>
+                ))
+              ) : (
+                <></>
+              )}
+            </Select>
+          </Grid> */}
           <Grid item xs={12} md={3}>
-            <InputLabel>Select Date:</InputLabel>
+            <InputLabel>Date:</InputLabel>
             <Select
               id="dateId"
               native
               onChange={handleChosenDateChange}
               fullWidth
               required
-              label="??"
               value={chosenDate}
             >
               <option aria-label="None" value="">
@@ -150,26 +173,46 @@ const LocationLog = (props) => {
       <div style={routeStyle}>
         Route {routeName} {routeDescription}
       </div>
-      <div>
-        {totalAttendeeCount} attendees moved in {entriesByDate.length} trips
-      </div>
-      <div>
-        <span>Time </span>
-        <span>Vehicle Number </span>
-        <span>Attendee Count </span>
-        <span>Entered By</span>
-      </div>
-      <div className="location_log_header"></div>
-      <div>
-        {entriesByDate.map((entry) => (
-          <EntryCard
-            key={entry.id}
-            entry={entry}
-            deleteThisEntry={deleteThisEntry}
-            {...props}
-          />
-        ))}
-      </div>
+      <Typography variant="h6">
+        {totalAttendeeCount} attendees moved in {entriesByDate.length} trips.
+      </Typography>
+      <TableContainer component={Paper}>
+        <Table size="small" aria-label="a dense table">
+          <TableHead>
+            <TableRow>
+              <TableCell>Time</TableCell>
+              <TableCell align="right">Vehicle #</TableCell>
+              <TableCell align="right">Attendee Count</TableCell>
+              <TableCell align="right">Entered By</TableCell>
+              <TableCell align="right">Actions</TableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {entriesByDate.map((entry) => (
+              <TableRow key={entry.id}>
+                <TableCell component="th" scope="entry">
+                  {entry.time}
+                </TableCell>
+                <TableCell align="right">{entry.vehicle_number}</TableCell>
+                <TableCell align="right">{entry.attendee_count}</TableCell>
+                <TableCell align="right">
+                  {entry.user.first_name} {entry.user.last_name}
+                </TableCell>
+                <TableCell align="right">
+                  <EditIcon
+                    onClick={() =>
+                      props.history.push(`/entry/edit/form/${entry.id}`)
+                    }
+                  />
+                  <DeleteOutlinedIcon
+                    onClick={() => deleteThisEntry(entry.id)}
+                  />
+                </TableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </TableContainer>
     </>
   );
 };
